moved reset and forgot endpoints to routes
